Fix decoding on Windows.
Because Windows expects crlf in files, and that might not be in the file
(indeed, it is not in our test files), we need to read raw bytes into Perl
first, and then decode them. So just prepend the :raw IO layer. Now the
UTF- tests pass on Windows withiut any "utf8 "\xA5" does not map to
"Unicode" warnings.